Basic minimax-style alpha-beta pruning.
Alpha beta does not directly integrate into expectiminimax due to CHANCE
nodes. This implementation therefore reassigns alpha to min and beta to
max for the children of chance nodes that have more than one child.

Also integrated into transposition tables as a bound. We can still use
entries that are known victories or loss, and skip entries with known
less work. Further, we can keep entries with exact scores and we can use
entries with bounded scores (from a previous cutoff) and compare to the
current alpha/beta in deciding whether that bounded score is sufficient.

Already we see some performance improvement in DiceBattle, due to the
small number of nodes that are deterministic.

Next up is *-minimax, which supplies alpha and beta in CHANCE nodes. And
then we may also do refutation tables and iterative deepening.