Two changes to chance node probing (one good, one bad, still WIP)
Changing chance node probing to require a depth > 1 is an overall smart
choice, as there is no way to establish a UB/LB, only to perform the
actual checking. If this isn't enough to cut off, the search will be
performed again -- a waste of time! Fixing this improved performance,
but not enough to make probing a good default.

Also fix a bug. Had intended probing to be used when branching factor is
high (supported by the slide deck I found the idea in), but had the
condition backwards (an experiment I accidentally committed?) and it
only ran on small nodes. "Fixing" this made probing worse.

We can't say that the change is "wrong" because it made things slower.
Basically, probing is currently slower, so any change that results in
doing less probing will appear to "help," but only so long as probing is
slow. If a better probing method is found that increases performance,
probing less is likely to hurt performance.

That said, only probing CHANCE nodes will small children n may make
sense. The argument I can make for this is that a cutoff will tend to be
late. If we cut off only the nth node, then we did n probes to save 1
full search. This 1/n supports higher effectiveness rates at lower
values of n.

If we can fix probing so that it's generally advantageous we'll be able
to answer these questions. Or, if we add a config, we can make this
a tunable parameter.