Add default mcts config to CLI tools constructor.
It's a little hacky. The max time setting should probably be removed
from config and changed to a parameter in chooseBest().

It also would be ideal if CLI tools took a default EngineConfig which
would allow users to specify that they wish to use xmm or mcts by
default. However, this won't play well with the way flags currently
work.