Rename project to DartNAD0: dartnad0's not alphazero its dartnad0
Full acronym is doubly recursive for: DartNAD0 A.I. for Randomized
Tabletop games is Not AlphaZero its DartNAD 0.

Reflects the generic nature of this engine, as well as the fact that we
support AlphaZero style PUCT, as well as the fact that this is most
definitely not AlphaZero, not least of all because it's written in Dart.