Create CliEngine class that encapsulates parsing engines on the cli.
Use this class in ParseEngineConfig in order to better encapsulate help
text and parsing and validation, etc, and be less error prone.

This also reduces the need to pass around defaultXmmConfig and
defaultMctsConfig everywhere. Soon to be made completely optional in the
CliTools constructor.

It should will be possible for users to specify their own CliEngines if
they wish.