Add a strict time mode which runs search on an isolate and kills at time
Seems to cause something like a 10-50 elo loss when enabled on
dicebattle with a mildly tuned buffer.

However, it may be useful in some cases, and it is useful for punishing
configs which perform slower than asked due to wild GC pauses.

Likely, to make the most of this feature, TimeController should be able
to roll time forward to the next move, and engines should stream their
results, so that *some* move is available even during GC pauses. And the
choice of killing isolates that go over time may be too harsh as well.